Add the correct include for display util and remerge
The display util file was moved out of chrome os on the master branch.
However that change is not merged to M67 branch. This patch fixes the
include that refers to the state in master.

It then re-merges
(merge) Replaces uint32_t with double when working with scales

Merge to M67
Zoom factors are double, but the code is using uint32_t to process them.
This patch fixes it.

Output on wayland info client:
OUTPUT0:

  connection:          internal
  device scale factor: 1.000

  geometry:
    x:                 2
    y:                 24
    physical width:    443 mm
    physical height:   264 mm
    subpixel:          unknown
    make:              unknown
    model:             unknown
    transform:         normal

  modes:
    1676x998:          60.00 Hz      current       preferred

  scales:
    0.500:
    0.750:
    1.000:             current       preferred
    1.250:
    1.500:
    1.750:
    2.000:
    2.250:
    2.500:
Component: Wayland, aura shell, scales
Reviewed-on: https://chromium-review.googlesource.com/1028217
Commit-Queue: Malay Keshav <malaykeshav@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#553586}
(cherry picked from commit e24cbe8a0ed6a7b9d96dc34560c8909ad75d2764)

Reviewed-on: https://chromium-review.googlesource.com/1030854
Cr-Commit-Position: refs/branch-heads/3396@{#334}
Cr-Branched-From: 9ef2aa869bc7bc0c089e255d698cca6e47d6b038-refs/heads/master@{#550428}

+++ components/exo/wayland/server.cc
@@ -95,7 +95,7 @@
 #include "ui/base/ui_features.h"
 #include "ui/compositor/compositor_vsync_manager.h"
 #include "ui/display/display_switches.h"
-#include "ui/display/manager/display_util.h"
+#include "ui/display/manager/chromeos/display_util.h"
 #include "ui/display/manager/managed_display_info.h"
 #include "ui/display/screen.h"
 #include "ui/events/keycodes/dom/keycode_converter.h"
@@ -2872,10 +2872,9 @@ class AuraOutput : public WaylandDisplayObserver::ScaleObserver {
         bool rv = display_manager->GetActiveModeForDisplayId(display.id(),
                                                              &active_mode);
         DCHECK(rv);
-        for (uint32_t zoom_factor :
-             display::GetDisplayZoomFactors(active_mode)) {
+        for (double zoom_factor : display::GetDisplayZoomFactors(active_mode)) {
           uint32_t flags = 0;
-          if (zoom_factor == 1.f)
+          if (zoom_factor == 1.0)
             flags |= ZAURA_OUTPUT_SCALE_PROPERTY_PREFERRED;
           if (display_info.zoom_factor() == zoom_factor)
             flags |= ZAURA_OUTPUT_SCALE_PROPERTY_CURRENT;
